MatGraphAI/requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
alabaster==0.7.13
anytree==2.8.0
asgiref==3.4.1
async-timeout==3.0.1
asyncio-pool==0.5.2
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.10.3
Beaker==1.11.0
bleach==5.0.1
blessings==1.7
blis==0.7.9
catalogue==2.0.8
certifi==2021.5.30
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.1
colorama==0.4.6
confection==0.0.4
constantly==15.1.0
cryptography==37.0.2
cryptojwt==1.8.2
cymem==2.0.7
Cython==0.29.34
daphne==3.0.2
defusedxml==0.7.1
dill==0.3.6
Django==3.2.16
django-admin-csvexport==1.11
django-admin-interface==0.19.1
django-admin-multiple-choice-list-filter==0.1.1
django-admin-relation-links==0.2.5
django-allauth==0.51.0
django-annoying==0.10.6
django-autocomplete-light==3.9.4
django-bulk-update-or-create==0.3.0
django-cleanup==5.2.0
django-colorfield==0.7.2
django-common-helpers==0.9.2
django-cors-headers==3.8.0
django-countries==7.2.1
django-cron==0.6.0
django-extensions==3.1.3
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-guardian==2.4.0
django-haystack==3.2.1
django-iban-field==0.8
django-ipware==5.0.0
django-localflavor==3.1
django-mailer==2.1
django-money==2.0.1
django-mysql==4.3.0
django-neomodel==0.0.7
django-nested-admin==3.4.0
django-nonrelated-inlines==0.1.1
django-picklefield==3.0.1
django-related-admin==1.0.1
django-request-logging==0.7.3
django-rest-auth==0.9.5
django-sortedm2m==3.0.2
django-sphinxdoc==2.0.0
django-summernote==0.8.20.0
djangorestframework==3.12.4
dnspython==1.16.0
docopt==0.6.2
docutils==0.18.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
eventlet==0.31.1
filelock==3.0.12
flatten-json==0.1.13
future==0.18.2
fuzzywuzzy==0.18.0
graphviz==0.20.1
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httptools==0.4.0
hyperlink==21.0.0
idna==3.2
imagesize==1.4.1
incremental==21.3.0
isodate==0.6.0
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
livereload==2.6.3
lockfile==0.12.2
lxml==4.6.5
Mako==1.1.5
MarkupSafe==2.0.1
msgpack==1.0.2
msgpack-python==0.5.6
multidict==5.1.0
murmurhash==1.0.9
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel @ git+https://github.com/neo4j-contrib/neomodel.git@2957f74a117f51dccf4b84129dc0b68a03e7bf6a
nltk==3.8.1
num2words==0.5.10
numpy==1.23.5
oauthlib==3.2.0
oic==1.3.0
oidcmsg==1.4.0
oidcservice==1.1.0
ontopy==0.1.1
openai==0.27.4
openpyxl==3.0.10
Owlready2==0.41
packaging==21.0
pandarallel==1.6.4
pandas==1.5.2
pathy==0.10.1
Pillow==9.1.1
preshed==3.0.8
psutil==5.9.5
psycopg2-binary==2.9.3
PubChemPy==1.0.4
py-moneyed==1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
pydantic==1.10.7
Pygments==2.15.0
PyHamcrest==2.0.2
pyjwkest==1.4.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyPDF2==1.27.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-monkey-business==1.0.0
python-stdnum==1.18
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
qrcode==7.3
qrplatba==0.4.1
rdflib==6.3.2
readerwriterlock==1.0.8
redis==3.5.3
regex==2023.3.23
reportlab==3.6.1
requests==2.28.0
requests-oauthlib==1.3.0
selenium==3.141.0
semver==2.13.0
sentry-sdk==1.10.1
service-identity==21.1.0
six==1.16.0
smart-open==6.3.0
snowballstemmer==2.2.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sparql-client==3.8
Sphinx==6.1.3
sphinx-autobuild==2021.3.14
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
srsly==2.4.6
stop-words==2018.7.23
stopwords==1.0.0
tabulate==0.8.9
tenacity==8.2.2
thinc==8.1.9
tornado==6.2
tqdm==4.65.0
Twisted==22.4.0
txaio==21.2.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.12
uvicorn==0.17.6
wasabi==1.1.1
webencodings==0.5.1
websockets==10.3
xlwt==1.3.0
yarl==1.7.2
zope.interface==5.4.0
